Added Auto-Installer method, reworked guide
Current features moved to bottom and is instead linked from "About" to reduce clutter, I feel it's more useful as a "changelog" kind of thing.
Added reference to getdotastats.com Lobby Browser
Toggle between skill sets should be inside FAQ section, maybe Ash47 can do this. It should not be at the top.

The rest is just structural logic. I tried to use as few words as possible to explain it in the easiest way I could think of